APIのフォルダ構成変えてみた．こっちのほうが，分類もできるし，ファイル探す手間は減ると思うけどどうでしょう．
